transfered tests to the right class
